Reward distribution. Types correction

+++ src/masternode-tiers.h
@@ -16,6 +16,6 @@ const double TIER_OBFUSCATION_VALUES[MasternodeTiers::TIER_NONE] = {999.99, 2999
 bool IsMasternodeOutput(const CWalletTx* tx, int index);
 int GetMasternodeTierFromOutput(COutput& out);
 double GetObfuscationValueForTier(int nTier);
-unsigned int CalculateWinningTier(std::vector<int>& vecTierSizes, uint256 blockHash);
+unsigned int CalculateWinningTier(std::vector<size_t>& vecTierSizes, uint256 blockHash);
 
 #endif
